+++ source/games/blood/src/nnexts.h
@@ -319,7 +319,7 @@ void damageSprites(DBloodActor* pXSource, DBloodActor* pSprite);
 void useRandomItemGen(DBloodActor* pSource);
 void useUniMissileGen(DBloodActor* sourceactor, DBloodActor* actor);
 void useSoundGen(DBloodActor* sourceactor, DBloodActor* actor);
-void useIncDecGen(DBloodActor* sourceactor, short objType, int objIndex, DBloodActor* objactor);
+void useIncDecGen(DBloodActor* sourceactor, int objType, int objIndex, DBloodActor* objactor);
 void useDataChanger(DBloodActor* sourceactor, int objType, int objIndex, DBloodActor* objActor);
 void useSectorLightChanger(DBloodActor* pXSource, sectortype* pSector);
 void useTargetChanger(DBloodActor* sourceactor, DBloodActor* actor);
